clean up comments

+++ contracts/market/OverlayV1OI.sol
@@ -52,64 +52,34 @@ contract OverlayV1Oi {
         uint112 fundingKDenominator,
         uint256 elapsed
     ) internal returns (uint256 amountToBurn) {
-
         FixedPoint.uq144x112 memory fundingFactor = computeFundingFactor(
             fundingKDenominator - 2 * fundingKNumerator,
             fundingKDenominator,
             elapsed
         );
 
-        // revert();
-
-        // if (oiShort == 0) {
-        //     uint256 oiLongNow = fundingFactor.mul(oiLong).decode144();
-        //     amountToBurn = oiLong - oiLongNow;
-        //     oiLong = oiLongNow;
-        // } else if (oiLong == 0) {
-        //     uint256 oiShortNow = fundingFactor.mul(oiShort).decode144();
-        //     amountToBurn = oiShort - oiShortNow;
-        //     oiShort = oiShortNow;
-        // } else if (oiLong > oiShort) {
-        //     uint256 oiImbNow = fundingFactor.mul(oiLong - oiShort).decode144();
-        //     oiLong = (oiLong + oiShort + oiImbNow) / 2;
-        //     oiShort = (oiLong + oiShort - oiImbNow) / 2;
-        // } else {
-        //     uint256 oiImbNow = fundingFactor.mul(oiShort - oiLong).decode144();
-        //     oiShort = (oiLong + oiShort + oiImbNow) / 2;
-        //     oiLong = (oiLong + oiShort - oiImbNow) / 2;
-        // }
-
-        // below version is 733 less expensive on gas and 171 bytes smaller
-        uint funding = oiLong;
-        uint funded = oiShort;
+        uint256 funding = oiLong;
+        uint256 funded = oiShort;
 
         bool paidByShorts = funding <= funded;
-
         if (paidByShorts) (funding, funded) = (funded, funding);
 
         if (funded == 0) {
-
-            uint oiNow = fundingFactor.mul(funding).decode144();
-
+            uint256 oiNow = fundingFactor.mul(funding).decode144();
             amountToBurn = funding - oiNow;
 
             if (paidByShorts) oiShort = oiNow;
             else oiLong = oiNow;
-
         } else {
-
-            uint oiImbNow = fundingFactor.mul(funding - funded).decode144();
-
-            uint total = funding + funded;
+            uint256 oiImbNow = fundingFactor.mul(funding - funded).decode144();
+            uint256 total = funding + funded;
 
             funding = ( total + oiImbNow ) / 2;
             funded = ( total - oiImbNow ) / 2;
 
             if (paidByShorts) ( oiShort = funding, oiLong = funded );
             else ( oiLong = funding, oiShort = funded );
-
         }
-
     }
 
     /// @notice Adds to queued open interest to prep for T+1 price settlement
